app/requirements.txt
fastapi
uvicorn
tensorflow==2.19.0
pillow
python-multipart
numpy
